requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cassandra-driver==3.28.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
geomet==0.2.1.post1
h11==0.14.0
idna==3.4
pydantic==2.4.2
pydantic-core==2.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
